Enables VirtualBox basic USB Controller
Software in the ODM collection of applications require a live USB controller or they will error. This change causes VirtualBox to enable its USB controller emulation feature. It does not enable the the VirtualBox USB 2.0 support as that is not required to get the ODM software to run without error.